+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ## [v6.0.0 (unreleased)](https://github.com/laravel/scout/compare/v5.0.3...v6.0.0)
 
+### Added
+- Method `count` to Builder ([#314](https://github.com/laravel/scout/pull/314))
+
 ### Changed
 - Flush records of a model using the engine. **This removes the emitting of the `ModelsFlushed` event.** ([#310](https://github.com/laravel/scout/pull/310))